+++ Core/HLE/sceKernelThread.h
@@ -17,6 +17,7 @@
 
 #pragma once
 
+#include <cstdio>
 #include <string>
 #include <vector>
 #include <map>
@@ -24,6 +25,7 @@
 
 #include "Common/CommonTypes.h"
 #include "Core/HLE/sceKernel.h"
+#include "Core/HLE/ErrorCodes.h"
 #include "Core/HLE/PSPThreadContext.h"
 #include "Core/HLE/KernelThreadDebugInterface.h"
 
@@ -173,6 +175,46 @@ struct NativeThread {
 	s32_le numInterruptPreempts;
 	s32_le numThreadPreempts;
 	s32_le numReleases;
+};
+
+struct NativeCallback {
+	SceUInt_le size;
+	char name[32];
+	SceUID_le threadId;
+	u32_le entrypoint;
+	u32_le commonArgument;
+
+	s32_le notifyCount;
+	s32_le notifyArg;
+};
+
+// Exposed here (rather than kept private to sceKernelThread.cpp) so the WebSocket debugger can
+// read a live object's state directly via kernelObjects.Get<PSPCallback>()/Iterate<PSPCallback>()
+// - see HLEKernelObjectSubscriber.cpp. That's a read-only use: nothing outside this file should
+// call DoState() or otherwise mutate a PSPCallback - it's public here for this file's own use as
+// before, not an invitation to write to it from elsewhere.
+class PSPCallback : public KernelObject {
+public:
+	const char *GetName() override { return nc.name; }
+	const char *GetTypeName() override { return GetStaticTypeName(); }
+	static const char *GetStaticTypeName() { return "CallBack"; }
+
+	void GetQuickInfo(char *ptr, int size) override {
+		snprintf(ptr, size, "thread=%i, argument= %08x",
+			nc.threadId,
+			nc.commonArgument);
+	}
+
+	~PSPCallback() {
+	}
+
+	static u32 GetMissingErrorCode() { return SCE_KERNEL_ERROR_UNKNOWN_CBID; }
+	static int GetStaticIDType() { return SCE_KERNEL_TMID_Callback; }
+	int GetIDType() const override { return SCE_KERNEL_TMID_Callback; }
+
+	void DoState(PointerWrap &p) override;
+
+	NativeCallback nc;
 };
 
 struct ThreadWaitInfo {
